lint: clean the pilot/research surfaces (274 -> 0), archive excluded
pyproject: extend-exclude act2-pilot-archive (retired vestige incl.
intentionally buggy sources). Mechanical, behavior-preserving fixes across
scripts/act2 (S1-S14 pilot scripts) + latent-gate:
- import sorting, unused imports/locals removed (all verified dead by ruff)
- int(round()) -> round() (py3 semantics identical)
- degenerate identical-branch conditionals collapsed (g1:110, s1:91, e3:46,
  train_embedder_yu:101) — behavior EXACTLY preserved, collapsed forms noted
- subprocess wrappers made explicit check=False (callers inspect rc)
- B023 loop-capture fixed properly (default-arg binding, not suppression)
- BLE001 noqa'd with reason on the two deliberate degrade-don't-die handlers
  (repo precedent: release_ceremony.py L114)
- open() calls moved to context managers
Full suite: 547 passed, 6 skipped; all touched scripts py_compile clean.

+++ packages/latent-gate/public/eval-pack/run_eval.py
@@ -73,7 +73,7 @@ def main() -> int:
     order = sorted(range(n), key=lambda i: -scored[i]["conf"])
     print(f"{'couverture':>10} {'n':>4} {'acc':>7} {'IC95':>16}")
     for cov in COVERAGES:
-        m = max(1, int(round(n * cov)))
+        m = max(1, round(n * cov))
         sel = order[:m]
         k = sum(1 for i in sel
                 if (scored[i]["p"] > 0.5) == bool(scored[i]["y"]))
